# .docs/IoT.TrafficLights.Api.Spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrafficLights.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Admin/ping": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Admin/config": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficControl/start": {
      "post": {
        "tags": [
          "TrafficControl"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficControl/stop": {
      "post": {
        "tags": [
          "TrafficControl"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficControl/standby": {
      "post": {
        "tags": [
          "TrafficControl"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficControl/shut": {
      "post": {
        "tags": [
          "TrafficControl"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficLog/id/{id}": {
      "get": {
        "tags": [
          "TrafficLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "TrafficLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TrafficLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficLog/date/{date}": {
      "get": {
        "tags": [
          "TrafficLog"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TrafficLog"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficLog/all": {
      "get": {
        "tags": [
          "TrafficLog"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TrafficLog"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TrafficLog": {
      "post": {
        "tags": [
          "TrafficLog"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TrafficLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "type": "string",
            "nullable": true
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}